refactor(#332): move event subscription logic into FileIndex.BuildAndWatch
server.go now just calls:
  idx.BuildAndWatch(deps.EventBus(), deps.StorageService().GetManagedDevices)

All device-resolution and event-dispatch logic lives in FileIndex itself,
co-located with the index operations it drives. Also extracts resolveDevice()
as a package-level helper for reuse.
